service/reader_requirement.txt
easyocr
pillow
flask
flask-cors
